Added points
Added a medium asteroid and tweaked movement so it feels much more responsive. Also made the ship smaller so it is less claustrophobic

Added points for destroying asteroids

Added a custom font which is based on the kind found in the original game. This font is used to display the current points