Significant refactor
Created a models class so it's easier to draw the lives

Removed homing torpedoes

Added game states; Start, Running, End. The game beings in the "Start" which plays a "demo" in the background. "Running" is the core game, and when the player runs out of lives it goes to "End" which just lets the player view their score before going back to the start. Spacebar is used to proceed

Huge refactor of the main update() function to account for the game state changes and to also try and fix index out of bounds issues where after I remove two bodies